Introduce functions to process profiles selections
This commits move all profile related rules from variables.py to a new
file called profiles.py and update the profile functions to make them
more generic and reusable. This also simplifies the variables.py files
which is now only used to process variables files, options and
respective values.
