+++ ssg/variables.py
@@ -4,16 +4,10 @@ from __future__ import print_function
 import glob
 import os
 import sys
-import yaml
 
 from collections import defaultdict
 from .constants import BENCHMARKS
-from .controls import ControlsManager
-from .products import (
-    get_profile_files_from_root,
-    load_product_yaml,
-    product_yaml_path,
-)
+from .profiles import get_profiles_from_products
 from .yaml import open_and_macro_expand
 
 
@@ -111,235 +105,12 @@ def get_variable_options(content_dir: str, variable_id: str = None) -> dict_type
     return all_options
 
 
-class ProfileVariables:
-    """
-    A class to represent profile variables.
-
-    Attributes:
-    -----------
-    profile_id : str
-        The unique identifier for the profile.
-    product : str
-        The product associated with the profile.
-    variables : dict
-        A dictionary containing the variables for the profile.
-    """
-    def __init__(self, profile_id, product, variables):
-        self.profile_id = profile_id
-        self.product = product
-        self.variables = variables
-
-
-def _load_product_yaml(content_dir: str, product: str) -> object:
-    """
-    Load the product YAML file and return its content as a Python object.
-
-    Args:
-        content_dir (str): The directory where the content is stored.
-        product (str): The name of the product.
-
-    Returns:
-        object: The loaded YAML content as a Python object.
-    """
-    file_yaml_path = product_yaml_path(content_dir, product)
-    return load_product_yaml(file_yaml_path)
-
-
-def _load_yaml_profile_file(file_path: str) -> dict_type:
-    """
-    Load the content of a YAML file intended to profiles definitions.
-
-    It is not necessary to process macros in this case.
-
-    Args:
-        file_path (str): The path to the YAML file.
-
-    Returns:
-        dict: The content of the YAML file as a dictionary.
-    """
-    with open(file_path, 'r') as file:
-        try:
-            return yaml.safe_load(file)
-        except yaml.YAMLError as e:
-            print(f"Error loading YAML profile file {file_path}: {e}")
-            return {}
-
-
-def _get_extended_profile_path(profiles_files: list, profile_name: str) -> str:
-    """
-    Retrieve the full path of a profile file from a list of profile file paths.
-
-    Args:
-        profiles_files (list of str): A list of file paths where profile files are located.
-        profile_name (str): The name of the profile to search for.
-
-    Returns:
-        str: The full path of the profile file if found, otherwise None.
-    """
-    profile_file = f"{profile_name}.profile"
-    profile_path = next((path for path in profiles_files if profile_file in path), None)
-    return profile_path
-
-
-def _process_profile_extension(profiles_files: list, profile_yaml: dict,
-                               profile_variables: dict, policies: dict) -> dict_type:
-    """
-    Processes the extension of a profile by recursively checking if the profile extends another
-    profile and updating the profile variables accordingly.
-
-    Args:
-        profiles_files (list): List of profile file paths.
-        profile_yaml (dict): The YAML content of the current profile.
-        profile_variables (dict): The variables already defined in the current profile.
-        policies (dict): The policies defined in the current profile.
-
-    Returns:
-        dict: The updated profile variables after processing the extended profile,
-              or the original profile variables if no extension is found.
-    """
-    extended_profile = profile_yaml.get("extends")
-    if isinstance(extended_profile, str):
-        extended_profile = _get_extended_profile_path(profiles_files, extended_profile)
-        if extended_profile is not None:
-            return _process_profile(profiles_files, extended_profile, policies, profile_variables)
-    return profile_variables
-
-
-def _process_controls(control_line: str, profile_variables: dict, policies: dict) -> dict_type:
-    """
-    Process a control file inheritance to update profile variables based on the given policies.
-
-    Args:
-        control_line (str): A string representing the control line, which contains a policy ID and
-                            optionally a level, separated by colons.
-        profile_variables (dict): A dictionary of profile variables to be updated.
-        policies (dict): A dictionary of policies, where each key is a policy ID and each value is
-                         a policy object containing the controls.
-
-    Returns:
-        dict: The updated profile variables dictionary.
-
-    Raises:
-        KeyError: If the policy ID from the control line is not found in the policies dictionary.
-    """
-    if control_line.count(":") == 2:
-        policy_id, _, level = control_line.split(":")
-    else:
-        policy_id, _ = control_line.split(":")
-        level = None
-
-    try:
-        policy = policies[policy_id]
-    except KeyError:
-        print(f"Policy {policy_id} not found")
-        return profile_variables
-
-    for control in policy.controls:
-        if level in control.levels:
-            for rule in control.rules:
-                if "=" in rule:
-                    variable_name, variable_value = rule.split('=', 1)
-                    # When a profile extends a control file, the variables explicitly defined in
-                    # profiles files must be honored, so don't update variables already defined.
-                    if variable_name not in profile_variables:
-                        profile_variables[variable_name] = variable_value
-    return profile_variables
-
-
-def _process_selections(profile_yaml: dict, profile_variables: dict, policies: dict) -> dict_type:
-    """
-    Processes the selections from the profile YAML and updates the profile variables accordingly.
-
-    Args:
-        profile_yaml (dict): A dictionary containing the profile YAML data.
-        profile_variables (dict): A dictionary to store the profile variables.
-        policies (dict): A dictionary containing policy information.
-
-    Returns:
-        dict: The updated profile variables dictionary.
-    """
-    selections = profile_yaml.get("selections")
-    for selected in selections:
-        if "=" in selected and "!" not in selected:
-            variable_name, variable_value = selected.split('=', 1)
-            profile_variables[variable_name] = variable_value
-        elif ":" in selected:
-            profile_variables = _process_controls(selected, profile_variables, policies)
-    return profile_variables
-
-
-def _process_profile(profiles_files: list, file: str, policies: dict,
-                     profile_variables={}) -> dict_type:
-    """
-    Processes a profile by loading its YAML file, handling profile extensions, and processing
-    selections.
-
-    Args:
-        profiles_files (list): A list of profile file paths.
-        file (str): The path to the profile file to be processed.
-        policies (dict): A dictionary of policies defined by control files.
-        profile_variables (dict, optional): A dictionary of profile variables. Defaults to empty.
-
-    Returns:
-        dict: A dictionary containing the processed profile variables.
-    """
-    profile_yaml = _load_yaml_profile_file(file)
-    profile_variables = _process_profile_extension(profiles_files, profile_yaml,
-                                                   profile_variables, policies)
-    profile_variables = _process_selections(profile_yaml, profile_variables, policies)
-    return profile_variables
-
-
-def _load_controls_manager(controls_dir: str, product_yaml: dict) -> object:
-    """
-    Loads and initializes a ControlsManager instance.
-
-    Args:
-        controls_dir (str): The directory containing control files.
-        product_yaml (dict): The product configuration in YAML format.
-
-    Returns:
-        object: An instance of ControlsManager with loaded controls.
-    """
-    control_mgr = ControlsManager(controls_dir, product_yaml)
-    control_mgr.load()
-    return control_mgr
-
-
-def _get_profiles_from_products(content_dir: str, products: list) -> list_type:
-    """
-    Retrieves profiles with respective variables from the given products.
-
-    Args:
-        content_dir (str): The directory containing the content.
-        products (list): A list of product names to retrieve profiles from.
-
-    Returns:
-        list: A list of ProfileVariables objects containing profile variables for each product.
-    """
-    profiles = []
-    controls_dir = os.path.join(content_dir, 'controls')
-
-    for product in products:
-        product_yaml = _load_product_yaml(content_dir, product)
-        profiles_files = get_profile_files_from_root(product_yaml, product_yaml)
-        controls_manager = _load_controls_manager(controls_dir, product_yaml)
-        for file in profiles_files:
-            profile_id = os.path.basename(file).split('.profile')[0]
-            profile_variables = _process_profile(profiles_files, file, controls_manager.policies)
-            profile = ProfileVariables(profile_id, product, profile_variables)
-            profiles.append(profile)
-
-    return profiles
-
-
 def _get_variables_from_profiles(profiles: list) -> dict_type:
     """
     Extracts variables from a list of profiles and organizes them into a nested dictionary.
 
     Args:
-        profiles (list): A list of profile objects, each containing variables, product, and id
-                         attributes.
+        profiles (list): A list of profile objects, each containing selections and id attributes.
 
     Returns:
         dict: A nested dictionary where the first level keys are variable names, the second level
@@ -383,7 +154,7 @@ def get_variables_by_products(content_dir: str, products: list) -> dict_type[str
         dict: A dictionary where keys are variable names and values are dictionaries of
               product-profile pairs.
     """
-    profiles = _get_profiles_from_products(content_dir, products)
+    profiles = get_profiles_from_products(content_dir, products)
     profiles_variables = _get_variables_from_profiles(profiles)
     return _convert_defaultdict_to_dict(profiles_variables)
